+++ memberaudit/managers/general.py
@@ -9,6 +9,7 @@ from esi.models import Token
 from eveuniverse.models import EveEntity, EveSolarSystem, EveType
 
 from allianceauth.services.hooks import get_extension_logger
+from app_utils.esi import fetch_esi_status
 from app_utils.logging import LoggerAddTag
 
 from .. import __title__
@@ -21,7 +22,6 @@ from ..constants import (
     EVE_CATEGORY_ID_SKILL,
     EVE_TYPE_ID_SOLAR_SYSTEM,
 )
-from ..helpers import fetch_esi_status
 from ..providers import esi
 
 logger = LoggerAddTag(get_extension_logger(__name__), __title__)