+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+### Changed
+
+- Use improved `fetch_esi_status()` from app_utils
+
+### Fixed
+
+- Trying to update a character on admin site gives error 500
+
 ## [1.3.2] - 2021-05-18
 
 ### Fixed